[estimator] clean up initialisation of joblib.Parallel
disabled n_jobs for windows, since it is broken for an
unknown reason.
