+++ pyemma/_ext/variational/solvers/eig_qr/.gitignore
@@ -1,0 +1,1 @@
+eig_qr.c
